+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+## TBD
+
+### Enhancements
+
+* Add PHP version string to report payload (device.runtimeVersions)
+  [#526](https://github.com/bugsnag/bugsnag-php/pull/526)
+
 ## 2.9.2 (2016-07-21)
 
 ### Bug Fixes